+++ hough3d/pointcloud.h
@@ -16,11 +16,11 @@
 #include <Eigen/Dense>
 
 // store points closer than dx to line (a, b) in Y
-Eigen::Matrix<double, 3, Eigen::Dynamic> pointsCloseToLine(const Eigen::Matrix<double, 3, Eigen::Dynamic>& X, const Eigen::Vector<double, 3> &a, const Eigen::Vector<double, 3> &b,  double dx);
-std::vector<Eigen::Index> indicesCloseToLine(const Eigen::Matrix<double, 3, Eigen::Dynamic>& Y, const Eigen::Vector<double,3> &a, const Eigen::Vector<double,3> &b, double dx);
+Eigen::Matrix<float, 3, Eigen::Dynamic> pointsCloseToLine(const Eigen::Matrix<float, 3, Eigen::Dynamic>& X, const Eigen::Vector<float, 3> &a, const Eigen::Vector<float, 3> &b,  float dx);
+std::vector<Eigen::Index> indicesCloseToLine(const Eigen::Matrix<float, 3, Eigen::Dynamic>& Y, const Eigen::Vector<float,3> &a, const Eigen::Vector<float,3> &b, float dx);
   
 // removes the points in Y from HoughPointCloud
 // WARNING: only works when points in same order as in HoughPointCloud!
-std::vector<Eigen::Index> removePoints(const Eigen::Matrix<double, 3, Eigen::Dynamic>& X, const Eigen::Matrix<double, 3, Eigen::Dynamic>& Y);
+std::vector<Eigen::Index> removePoints(const Eigen::Matrix<float, 3, Eigen::Dynamic>& X, const Eigen::Matrix<float, 3, Eigen::Dynamic>& Y);
 
 #endif /* HoughPointCloud_H_ */
